Move entrypoint back to gui module
