+++ pyproject.toml
@@ -37,7 +37,7 @@ dependencies = [
 Repository = "https://github.com/bohning/usdb_syncer"
 
 [project.scripts]
-usdb_syncer = "usdb_syncer.main:main"
+usdb_syncer = "usdb_syncer.gui.main:main"
 
 [build-system]
 requires = ["hatchling", "hatch-vcs"]
